move data fetching to CardSection component
