+++ src/Components/Common/CardSection.jsx
@@ -1,9 +1,28 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { Grid, Box, Typography } from '@mui/material';
 import PropTypes from 'prop-types';
 import PlanCard from './PlanCard';
+import { getData } from '../../util';
 
 const CardSection = ({ title, category='', search='' }) => {
+  const [plans, setPlans] = useState([]);
+  const URL = `${import.meta.env.VITE_API_BASE_URL}/api/v1/plans?categoryId=${category}&search=${search}&size=4`;
+
+  // TODO: Add skeleton loading feature later
+
+  useEffect(() => {
+    const fetchAllPlans = async () => {
+      try {
+        const res = await getData(URL);
+        setPlans(res?.items || []);
+        console.log(res);
+      } catch (error) {
+        console.log("Error fetching data:", error);
+      }
+    };
+    fetchAllPlans();
+  }, []);
+
   return (
     <Box
       sx={{
@@ -23,9 +42,9 @@ const CardSection = ({ title, category='', search='' }) => {
         }}
       >
         <Grid container spacing={3}>
-          {Array.from({ length: 4 }).map((_, index) => (
-            <Grid size={{ xs: 12, sm: 6, md: 3 }} key={index}>
-              <PlanCard planId={index} image='http://something' title='Placeholder example plan title here' />
+          {plans.map((plan) => (
+            <Grid size={{ xs: 12, sm: 6, md: 3 }} key={plan._id}>
+              <PlanCard {...plan} planId={plan._id} image={plan.images[0]} />
             </Grid>
           ))}
         </Grid>
